temp_backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-MySQLdb==2.0.0
Flask-JWT-Extended==4.6.0
mysqlclient==2.2.1
Werkzeug==3.0.1
python-dotenv==1.0.0
pydantic==2.5.3
pydantic[email]==2.5.3
email-validator==2.1.0
python-dateutil==2.8.2
